Deliver Global Labels within outline for P1 now too
