+++ spin2/src/spin1.semanticAndOutline.ts
@@ -4,6 +4,21 @@
 import * as vscode from "vscode";
 //import { Z_UNKNOWN } from "zlib";
 import { semanticConfiguration, reloadSemanticConfiguration } from "./spin2.semantic.configuration";
+import { ParseUtils } from "./spin1.utils";
+
+enum eParseState {
+  Unknown = 0,
+  inCon,
+  inDat,
+  inObj,
+  inPub,
+  inPri,
+  inVar,
+  inDatPasm,
+  inMultiLineComment,
+  inMultiLineDocComment,
+  inNothing,
+}
 
 // ----------------------------------------------------------------------------
 //  this file contains both an outline provider
@@ -14,16 +29,41 @@ import { semanticConfiguration, reloadSemanticConfiguration } from "./spin2.sema
 //   OUTLINE Provider
 //
 export class Spin1ConfigDocumentSymbolProvider implements vscode.DocumentSymbolProvider {
+  private parseUtils = new ParseUtils();
+  public constructor() {
+    if (this.spin1OutlineLogEnabled) {
+      if (this.spin1OutlineLog === undefined) {
+        //Create output channel
+        this.spin1OutlineLog = vscode.window.createOutputChannel("Spin1 Outline DEBUG");
+        this._logMessage("Spin1 Outline log started.");
+      } else {
+        this._logMessage("\n\n------------------   NEW FILE ----------------\n\n");
+      }
+    }
+  }
   public provideDocumentSymbols(document: vscode.TextDocument, _token: vscode.CancellationToken): Promise<vscode.DocumentSymbol[]> {
     return new Promise((resolve, _reject) => {
       let symbols: vscode.DocumentSymbol[] = [];
 
+      let currState: eParseState = eParseState.inCon; // compiler defaults to CON at start!
+      let priorState: eParseState = currState;
+      let prePasmState: eParseState = currState;
+
       for (let i = 0; i < document.lineCount; i++) {
         let line = document.lineAt(i);
+        const trimmedLine = line.text.trim();
+        const trimmedNonCommentLine = this.parseUtils.getNonCommentLineRemainder(0, line.text);
+
         let linePrefix: string = line.text;
         let lineHasComment: boolean = false;
         let commentOffset: number = 0;
         let commentLength: number = 0;
+
+        const sectionStatus = this._isSectionStartLine(line.text);
+        if (sectionStatus.isSectionStart) {
+          currState = sectionStatus.inProgressStatus;
+        }
+
         if (line.text.length > 2) {
           const lineParts: string[] = linePrefix.split(/[ \t]/);
           linePrefix = lineParts.length > 0 ? lineParts[0].toUpperCase() : "";
@@ -45,38 +85,209 @@ export class Spin1ConfigDocumentSymbolProvider implements vscode.DocumentSymbolP
           }
         }
 
-        if (linePrefix == "CON" || linePrefix == "DAT" || linePrefix == "VAR" || linePrefix == "OBJ") {
-          let sectionComment = lineHasComment ? line.text.substr(commentOffset, commentLength) : "";
-          const marker_symbol = new vscode.DocumentSymbol(linePrefix + " " + sectionComment, "", vscode.SymbolKind.Field, line.range, line.range);
+        if (sectionStatus.isSectionStart) {
+          if (linePrefix == "CON" || linePrefix == "DAT" || linePrefix == "VAR" || linePrefix == "OBJ") {
+            // start CON/VAR/OBJ/DAT
+            let sectionComment = lineHasComment ? line.text.substr(commentOffset, commentLength) : "";
+            const marker_symbol = new vscode.DocumentSymbol(linePrefix + " " + sectionComment, "", vscode.SymbolKind.Field, line.range, line.range);
 
-          symbols.push(marker_symbol);
-        } else if (linePrefix == "PUB" || linePrefix == "PRI") {
-          let methodScope: string = "Public";
-          if (line.text.startsWith("PRI")) {
-            methodScope = "Private";
-          }
-          let methodName: string = line.text.substr(3).trim();
-          if (methodName.includes("'")) {
-            const lineParts: string[] = methodName.split("'");
-            methodName = lineParts[0].trim();
-          }
-          if (methodName.includes("{")) {
-            const lineParts: string[] = methodName.split("{");
-            methodName = lineParts[0].trim();
-          }
-          if (methodName.includes("|")) {
-            const lineParts: string[] = methodName.split("|");
-            methodName = lineParts[0].trim();
-          }
+            symbols.push(marker_symbol);
+          } else if (linePrefix == "PUB" || linePrefix == "PRI") {
+            // start PUB/PRI
+            let methodScope: string = "Public";
+            if (line.text.startsWith("PRI")) {
+              methodScope = "Private";
+            }
+            let methodName: string = line.text.substr(3).trim();
+            if (methodName.includes("'")) {
+              const lineParts: string[] = methodName.split("'");
+              methodName = lineParts[0].trim();
+            }
+            if (methodName.includes("{")) {
+              const lineParts: string[] = methodName.split("{");
+              methodName = lineParts[0].trim();
+            }
+            if (methodName.includes("|")) {
+              const lineParts: string[] = methodName.split("|");
+              methodName = lineParts[0].trim();
+            }
 
-          const cmd_symbol = new vscode.DocumentSymbol(linePrefix + " " + methodName, "", vscode.SymbolKind.Function, line.range, line.range);
+            // NOTE this changed to METHOD when we added global labels which are to be Functions!
+            const cmd_symbol = new vscode.DocumentSymbol(linePrefix + " " + methodName, "", vscode.SymbolKind.Method, line.range, line.range);
 
-          symbols.push(cmd_symbol);
+            symbols.push(cmd_symbol);
+          }
+        } else {
+          let global_label: string | undefined = undefined;
+          if (trimmedNonCommentLine.length > 0) {
+            //this._logMessage("  * [" + currState + "] ln:" + (i + 1) + " trimmedNonCommentLine=[" + trimmedNonCommentLine + "]");
+            // NOT a section start
+            if (currState == eParseState.inDatPasm) {
+              // process pasm (assembly) lines
+              if (trimmedLine.length > 0) {
+                this._logMessage("    scan inDatPasm ln:" + (i + 1) + " trimmedNonCommentLine=[" + trimmedNonCommentLine + "]");
+                const lineParts: string[] = trimmedNonCommentLine.split(/[ \t]/);
+                if (lineParts.length > 0 && lineParts[0].toUpperCase() == "FIT") {
+                  this._logMessage("  - (" + (i + 1) + "): pre-scan DAT PASM line trimmedLine=[" + trimmedLine + "]");
+                  currState = prePasmState;
+                  this._logMessage("    scan END DATPasm ln:" + (i + 1) + " POP currState=[" + currState + "]");
+                  // and ignore rest of this line
+                  continue;
+                }
+                // didn't leave this state check for new global label
+                global_label = this._getDAT_PasmDeclaration(0, line.text); // let's get possible label on this ORG statement
+              }
+            } else if (currState == eParseState.inDat) {
+              this._logMessage("    scan inDat ln:" + (i + 1) + " trimmedNonCommentLine=[" + trimmedNonCommentLine + "]");
+              if (trimmedNonCommentLine.length > 6 && trimmedNonCommentLine.toUpperCase().includes("ORG")) {
+                // ORG, ORGF, ORGH
+                const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(trimmedNonCommentLine);
+                if (nonStringLine.toUpperCase().includes("ORG")) {
+                  this._logMessage("  - pre-scan DAT line trimmedLine=[" + trimmedLine + "] now Dat PASM");
+                  prePasmState = currState;
+                  currState = eParseState.inDatPasm;
+                  this._logMessage("    scan START DATPasm ln:" + (i + 1) + " PUSH currState=[" + prePasmState + "]");
+                  // and ignore rest of this line
+                  global_label = this._getDAT_PasmDeclaration(0, line.text); // let's get possible label on this ORG statement
+                }
+              } else {
+                global_label = this._getDAT_Declaration(0, line.text);
+              }
+            }
+            if (global_label != undefined) {
+              // was Variable: sorta OK (image good, color bad)
+              // was Constant: sorta OK (image good, color bad)   SAME
+              const global_symbol = new vscode.DocumentSymbol(global_label, "", vscode.SymbolKind.Constant, line.range, line.range);
+
+              symbols.push(global_symbol);
+            }
+          }
         }
       }
 
       resolve(symbols);
     });
+  }
+  private spin1OutlineLogEnabled: boolean = false; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit
+  private spin1OutlineLog: any = undefined;
+
+  private _logMessage(message: string): void {
+    if (this.spin1OutlineLog != undefined) {
+      //Write to output window.
+      this.spin1OutlineLog.appendLine(message);
+    }
+  }
+
+  private _isSectionStartLine(line: string): {
+    isSectionStart: boolean;
+    inProgressStatus: eParseState;
+  } {
+    // return T/F where T means our string starts a new section!
+    let startStatus: boolean = false;
+    let inProgressState: eParseState = eParseState.Unknown;
+    if (line.length > 2) {
+      const lineParts: string[] = line.split(/[ \t]/);
+      if (lineParts.length > 0) {
+        const sectionName: string = lineParts[0].toUpperCase();
+        startStatus = true;
+        if (sectionName === "CON") {
+          inProgressState = eParseState.inCon;
+        } else if (sectionName === "DAT") {
+          inProgressState = eParseState.inDat;
+        } else if (sectionName === "OBJ") {
+          inProgressState = eParseState.inObj;
+        } else if (sectionName === "PUB") {
+          inProgressState = eParseState.inPub;
+        } else if (sectionName === "PRI") {
+          inProgressState = eParseState.inPri;
+        } else if (sectionName === "VAR") {
+          inProgressState = eParseState.inVar;
+        } else {
+          startStatus = false;
+        }
+      }
+    }
+    if (startStatus) {
+      this._logMessage("** isSectStart line=[" + line + "]");
+    }
+    return {
+      isSectionStart: startStatus,
+      inProgressStatus: inProgressState,
+    };
+  }
+
+  private _getDAT_Declaration(startingOffset: number, line: string): string | undefined {
+    // HAVE    bGammaEnable        BYTE   TRUE               ' comment
+    //         didShow             byte   FALSE[256]
+    //                             byte   FALSE[256]
+    let newGlobalLabel: string | undefined = undefined;
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    // get line parts - we only care about first one
+    const dataDeclNonCommentStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    let lineParts: string[] = this.parseUtils.getNonWhiteLineParts(dataDeclNonCommentStr);
+    this._logMessage("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ")");
+    let haveMoreThanDat: boolean = lineParts.length > 1 && lineParts[0].toUpperCase() == "DAT";
+    if (haveMoreThanDat || lineParts[0].toUpperCase() != "DAT") {
+      // remember this object name so we can annotate a call to it
+      let nameIndex: number = 0;
+      let typeIndex: number = 1;
+      let maxParts: number = 2;
+      if (lineParts[0].toUpperCase() == "DAT") {
+        nameIndex = 1;
+        typeIndex = 2;
+        maxParts = 3;
+      }
+      let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[nameIndex]);
+      const isDataDeclarationLine: boolean = lineParts.length > maxParts - 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[typeIndex]) ? true : false;
+      let lblFlag: string = haveLabel ? "T" : "F";
+      let dataDeclFlag: string = isDataDeclarationLine ? "T" : "F";
+      this._logMessage("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ") label=" + lblFlag + ", daDecl=" + dataDeclFlag);
+      if (haveLabel) {
+        let newName = lineParts[nameIndex];
+        if (
+          !newName.toLowerCase().startsWith("debug") &&
+          !this.parseUtils.isPasmReservedWord(newName) &&
+          !this.parseUtils.isSpinBuiltInVariable(newName) &&
+          !this.parseUtils.isBuiltinReservedWord(newName)
+        ) {
+          if (!isDataDeclarationLine && !newName.startsWith(".") && !newName.startsWith(":") && !newName.includes("#")) {
+            newGlobalLabel = newName;
+          }
+          this._logMessage("  -- GLBL gddcl newName=[" + newGlobalLabel + "]");
+          //this._setGlobalToken(newName, new RememberedToken(nameType, labelModifiers));
+        }
+      }
+    }
+    return newGlobalLabel;
+  }
+
+  private _getDAT_PasmDeclaration(startingOffset: number, line: string): string | undefined {
+    // HAVE    bGammaEnable        BYTE   TRUE               ' comment
+    //         didShow             byte   FALSE[256]
+    let newGlobalLabel: string | undefined = undefined;
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    // get line parts - we only care about first one
+    const datPasmRHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    const lineParts: string[] = this.parseUtils.getNonWhiteLineParts(datPasmRHSStr);
+    //this._logPASM('- GetDATPasmDecl lineParts=[' + lineParts + ']');
+    // handle name in 1 column
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
+    if (haveLabel && !isDataDeclarationLine && !lineParts[0].startsWith(".") && !lineParts[0].startsWith(":") && !lineParts[0].includes("#")) {
+      const labelName: string = lineParts[0];
+      if (
+        !this.parseUtils.isReservedPasmSymbols(labelName) &&
+        !labelName.toUpperCase().startsWith("IF_") &&
+        !labelName.toUpperCase().startsWith("_RET_") &&
+        !labelName.toUpperCase().startsWith("DEBUG")
+      ) {
+        // org in first column is not label name, nor is if_ conditional
+        newGlobalLabel = labelName;
+        this._logMessage("  -- DAT PASM GLBL newGlobalLabel=[" + newGlobalLabel + "]");
+        //this._setGlobalToken(labelName, new RememberedToken(labelType, labelModifiers));
+      }
+    }
+    return newGlobalLabel;
   }
 }
 
@@ -182,21 +393,9 @@ interface IFilteredStrings {
   lineParts: string[];
 }
 
-enum eParseState {
-  Unknown = 0,
-  inCon,
-  inDat,
-  inObj,
-  inPub,
-  inPri,
-  inVar,
-  inDatPasm,
-  inMultiLineComment,
-  inMultiLineDocComment,
-  inNothing,
-}
-
 export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSemanticTokensProvider {
+  private parseUtils = new ParseUtils();
+
   async provideDocumentSemanticTokens(document: vscode.TextDocument, cancelToken: vscode.CancellationToken): Promise<vscode.SemanticTokens> {
     // SEE https://www.codota.com/code/javascript/functions/vscode/CancellationToken/isCancellationRequested
     if (cancelToken) {
@@ -286,7 +485,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     for (let i = 0; i < lines.length; i++) {
       const line = lines[i];
       const trimmedLine = line.trim();
-      const trimmedNonCommentLine = this._getNonCommentLineRemainder(0, line);
+      const trimmedNonCommentLine = this.parseUtils.getNonCommentLineRemainder(0, line);
       const sectionStatus = this._isSectionStartLine(line);
       const lineParts: string[] = trimmedNonCommentLine.split(/[ \t]/);
       if (currState == eParseState.inMultiLineComment) {
@@ -307,7 +506,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         }
         //  DO NOTHING Let Syntax highlighting do this
         continue;
-      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+      } else if (this.parseUtils.isFlexspinPreprocessorDirective(lineParts[0])) {
         this._getPreProcessor_Declaration(0, i + 1, line);
         continue;
       } else if (sectionStatus.isSectionStart) {
@@ -328,7 +527,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           // process a class(static) variable line
           if (trimmedNonCommentLine.length > 6 && trimmedNonCommentLine.toUpperCase().includes("ORG")) {
             // ORG, ORGF, ORGH
-            const nonStringLine: string = this._removeDoubleQuotedStrings(trimmedNonCommentLine);
+            const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(trimmedNonCommentLine);
             if (nonStringLine.toUpperCase().includes("ORG")) {
               this._logPASM("- (" + (i + 1) + "): pre-scan DAT line trimmedLine=[" + trimmedLine + "] now Dat PASM");
               prePasmState = currState;
@@ -390,7 +589,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           if (trimmedLine.length > 6) {
             if (trimmedLine.toUpperCase().includes("ORG")) {
               // ORG, ORGF, ORGH
-              const nonStringLine: string = this._removeDoubleQuotedStrings(trimmedLine);
+              const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(trimmedLine);
               if (nonStringLine.toUpperCase().includes("ORG")) {
                 this._logPASM("- (" + (i + 1) + "): pre-scan DAT line trimmedLine=[" + trimmedLine + "] now Dat PASM");
                 prePasmState = currState;
@@ -468,7 +667,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           currState = priorState;
         }
         //  DO NOTHING Let Syntax highlighting do this
-      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+      } else if (this.parseUtils.isFlexspinPreprocessorDirective(lineParts[0])) {
         const partialTokenSet: IParsedToken[] = this._reportPreProcessorLine(i, 0, line);
         partialTokenSet.forEach((newToken) => {
           this._logPreProc("=> PreProc: " + this._tokenString(newToken, line));
@@ -504,7 +703,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           // process a possible constant use on the CON line itself!
           if (line.length > 3) {
             if (trimmedLine.length > 6) {
-              const nonCommentLineRemainder: string = this._getNonCommentLineRemainder(0, trimmedLine);
+              const nonCommentLineRemainder: string = this.parseUtils.getNonCommentLineRemainder(0, trimmedLine);
               let orgStr: string = "ORGH";
               let orgOffset: number = nonCommentLineRemainder.toUpperCase().indexOf(orgStr); // ORGH
               if (orgOffset == -1) {
@@ -517,7 +716,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               }
               if (orgOffset != -1) {
                 // let's double check this is NOT in quoted string
-                const nonStringLine: string = this._removeDoubleQuotedStrings(nonCommentLineRemainder);
+                const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(nonCommentLineRemainder);
                 orgOffset = nonStringLine.toUpperCase().indexOf(orgStr); // ORG, ORGF, ORGH
               }
               if (orgOffset != -1) {
@@ -607,7 +806,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         // process a line in a data section
         if (trimmedLine.length > 0) {
           this._logDAT("- process DAT line(" + (i + 1) + "): trimmedLine=[" + trimmedLine + "]");
-          const nonCommentLineRemainder: string = this._getNonCommentLineRemainder(0, trimmedLine);
+          const nonCommentLineRemainder: string = this.parseUtils.getNonCommentLineRemainder(0, trimmedLine);
           let orgStr: string = "ORGH";
           let orgOffset: number = nonCommentLineRemainder.toUpperCase().indexOf(orgStr); // ORGH
           if (orgOffset == -1) {
@@ -620,7 +819,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           }
           if (orgOffset != -1) {
             // let's double check this is NOT in quoted string
-            const nonStringLine: string = this._removeDoubleQuotedStrings(nonCommentLineRemainder);
+            const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(nonCommentLineRemainder);
             orgOffset = nonStringLine.toUpperCase().indexOf(orgStr); // ORG, ORGF, ORGH
           }
           if (orgOffset != -1) {
@@ -698,14 +897,14 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
 
   private _getPreProcessor_Declaration(startingOffset: number, lineNbr: number, line: string): void {
     if (this.configuration.highlightFlexspin) {
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+      let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+      const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
       // get line parts - we only care about first one
       const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
       this._logPreProc("  - ln:" + lineNbr + " GetPreProcDecl lineParts=[" + lineParts + "]");
       const directive: string = lineParts[0];
       const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
-      if (this._isFlexspinPreprocessorDirective(directive)) {
+      if (this.parseUtils.isFlexspinPreprocessorDirective(directive)) {
         // check a valid preprocessor line for a declaration
         if (symbolName != undefined && directive.toLowerCase() == "#define") {
           this._logPreProc("  -- new PreProc Symbol=[" + symbolName + "]");
@@ -721,8 +920,8 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     //
     if (line.substr(startingOffset).length > 1) {
       //skip Past Whitespace
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+      let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+      const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
       this._logCON("  - ln:" + lineNbr + " GetCONDecl nonCommentConstantLine=[" + nonCommentConstantLine + "]");
 
       const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
@@ -773,10 +972,10 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // HAVE    bGammaEnable        BYTE   TRUE               ' comment
     //         didShow             byte   FALSE[256]
     //                             byte   FALSE[256]
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const dataDeclNonCommentStr = this._getNonCommentLineRemainder(currentOffset, line);
-    let lineParts: string[] = this._getNonWhiteLineParts(dataDeclNonCommentStr);
+    const dataDeclNonCommentStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    let lineParts: string[] = this.parseUtils.getNonWhiteLineParts(dataDeclNonCommentStr);
     this._logDAT("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ")");
     let haveMoreThanDat: boolean = lineParts.length > 1 && lineParts[0].toUpperCase() == "DAT";
     if (haveMoreThanDat || lineParts[0].toUpperCase() != "DAT") {
@@ -789,14 +988,14 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         typeIndex = 2;
         maxParts = 3;
       }
-      let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[nameIndex]);
-      const isDataDeclarationLine: boolean = lineParts.length > maxParts - 1 && haveLabel && this._isDatStorageType(lineParts[typeIndex]) ? true : false;
+      let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[nameIndex]);
+      const isDataDeclarationLine: boolean = lineParts.length > maxParts - 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[typeIndex]) ? true : false;
       let lblFlag: string = haveLabel ? "T" : "F";
       let dataDeclFlag: string = isDataDeclarationLine ? "T" : "F";
       this._logDAT("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ") label=" + lblFlag + ", daDecl=" + dataDeclFlag);
       if (haveLabel) {
         let newName = lineParts[nameIndex];
-        if (!this._isSpin2ReservedWords(newName)) {
+        if (!this.parseUtils.isSpin2ReservedWords(newName)) {
           const nameType: string = isDataDeclarationLine ? "variable" : "label";
           var labelModifiers: string[] = [];
           if (!isDataDeclarationLine) {
@@ -816,17 +1015,17 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _getDAT_PasmDeclaration(startingOffset: number, line: string): void {
     // HAVE    bGammaEnable        BYTE   TRUE               ' comment
     //         didShow             byte   FALSE[256]
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const datPasmRHSStr = this._getNonCommentLineRemainder(currentOffset, line);
-    const lineParts: string[] = this._getNonWhiteLineParts(datPasmRHSStr);
+    const datPasmRHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    const lineParts: string[] = this.parseUtils.getNonWhiteLineParts(datPasmRHSStr);
     //this._logPASM('- GetDATPasmDecl lineParts=[' + lineParts + ']');
     // handle name in 1 column
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
-      if (!this._isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_")) {
+      if (!this.parseUtils.isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_")) {
         // org in first column is not label name, nor is if_ conditional
         const labelType: string = isDataDeclarationLine ? "variable" : "label";
         var labelModifiers: string[] = [];
@@ -844,8 +1043,8 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     //  -or-   segments[7]     : "isp_hub75_segment"
     //
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const remainingNonCommentLineStr: string = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const remainingNonCommentLineStr: string = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     //this._logOBJ('- RptObjDecl remainingNonCommentLineStr=[' + remainingNonCommentLineStr + ']');
     const remainingLength: number = remainingNonCommentLineStr.length;
     if (remainingLength > 0) {
@@ -863,7 +1062,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // reset our list of local variables
     const isPrivate = methodType.indexOf("PRI") != -1;
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const startNameOffset = currentOffset;
     // find open paren
     currentOffset = line.indexOf("(", startNameOffset); // in spin1 ()'s are optional!
@@ -895,14 +1094,14 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _getSPIN_PasmDeclaration(startingOffset: number, line: string): void {
     // HAVE    next8SLine ' or .nextLine in col 0
     //         nPhysLineIdx        long    0
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const inLinePasmRHSStr = this._getNonCommentLineRemainder(currentOffset, line);
-    const lineParts: string[] = this._getNonWhiteLineParts(inLinePasmRHSStr);
+    const inLinePasmRHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    const lineParts: string[] = this.parseUtils.getNonWhiteLineParts(inLinePasmRHSStr);
     //this._logPASM('- GetInLinePasmDecl lineParts=[' + lineParts + ']');
     // handle name in 1 column
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
       const labelType: string = isDataDeclarationLine ? "variable" : "label";
@@ -919,12 +1118,12 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // HAVE    long    demoPausePeriod   ' comment
     //
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const remainingNonCommentLineStr: string = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const remainingNonCommentLineStr: string = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     this._logVAR("- GetVarDecl remainingNonCommentLineStr=[" + remainingNonCommentLineStr + "]");
     const isMultiDeclaration: boolean = remainingNonCommentLineStr.includes(",");
-    let lineParts: string[] = this._getNonWhiteDataInitLineParts(remainingNonCommentLineStr);
-    const hasGoodType: boolean = this._isStorageType(lineParts[0]);
+    let lineParts: string[] = this.parseUtils.getNonWhiteDataInitLineParts(remainingNonCommentLineStr);
+    const hasGoodType: boolean = this.parseUtils.isStorageType(lineParts[0]);
     this._logVAR("  -- lineParts=[" + lineParts + "]");
     let nameSet: string[] = [];
     if (hasGoodType && lineParts.length > 1) {
@@ -960,15 +1159,15 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportPreProcessorLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     // get line parts - we only care about first one
     const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
     this._logPreProc("  - ln:" + lineNumber + " reportPreProc lineParts=[" + lineParts + "]");
     const directive: string = lineParts[0];
     const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
     if (this.configuration.highlightFlexspin) {
-      if (this._isFlexspinPreprocessorDirective(directive)) {
+      if (this.parseUtils.isFlexspinPreprocessorDirective(directive)) {
         // record the directive
         tokenSet.push({
           line: lineNumber,
@@ -1001,7 +1200,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               ptTokenType: referenceDetails.type,
               ptTokenModifiers: updatedModificationSet,
             });
-          } else if (this._isFlexspinReservedWord(symbolName)) {
+          } else if (this.parseUtils.isFlexspinReservedWord(symbolName)) {
             // record a constant reference
             tokenSet.push({
               line: lineNumber,
@@ -1041,7 +1240,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportCON_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const nonCommentConstantLine = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
 
     const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
@@ -1097,7 +1296,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
           const assignmentRHSStr = conDeclarationLine.substring(fistEqualOffset + 1).trim();
           currentOffset = line.indexOf(assignmentRHSStr); // skip to RHS of assignment
           this._logCON("  -- CON assignmentRHSStr=[" + assignmentRHSStr + "]");
-          const possNames: string[] = this._getNonWhiteCONLineParts(assignmentRHSStr);
+          const possNames: string[] = this.parseUtils.getNonWhiteCONLineParts(assignmentRHSStr);
           this._logCON("  -- possNames=[" + possNames + "]");
           for (let index = 0; index < possNames.length; index++) {
             const possibleName = possNames[index];
@@ -1140,11 +1339,11 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 });
               } else {
                 if (
-                  !this._isSpinReservedWord(namePart) &&
-                  !this._isBuiltinReservedWord(namePart) &&
-                  !this._isUnaryOperator(namePart) &&
-                  !this._isSpinBuiltInConstant(namePart) &&
-                  !this._isPasmReservedWord(namePart)
+                  !this.parseUtils.isSpinReservedWord(namePart) &&
+                  !this.parseUtils.isBuiltinReservedWord(namePart) &&
+                  !this.parseUtils.isUnaryOperator(namePart) &&
+                  !this.parseUtils.isSpinBuiltInConstant(namePart) &&
+                  !this.parseUtils.isPasmReservedWord(namePart)
                 ) {
                   this._logCON("  --  CON MISSING name=[" + namePart + "]");
                   tokenSet.push({
@@ -1230,14 +1429,14 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportDAT_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const dataDeclNonCommentStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    let lineParts: string[] = this._getNonWhiteLineParts(dataDeclNonCommentStr);
+    let lineParts: string[] = this.parseUtils.getNonWhiteLineParts(dataDeclNonCommentStr);
     this._logVAR("- rptDataDeclLn lineParts=[" + lineParts + "]");
     // remember this object name so we can annotate a call to it
     if (lineParts.length > 0) {
-      if (this._isStorageType(lineParts[0]) || lineParts[0].toUpperCase() == "FILE" || lineParts[0].toUpperCase() == "ORG") {
+      if (this.parseUtils.isStorageType(lineParts[0]) || lineParts[0].toUpperCase() == "FILE" || lineParts[0].toUpperCase() == "ORG") {
         // if we start with storage type (or FILE, or ORG), not name, process rest of line for symbols
         currentOffset = line.indexOf(lineParts[0], currentOffset);
         const allowLocalVarStatus: boolean = false;
@@ -1265,10 +1464,10 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             ptTokenModifiers: referenceDetails.modifiers,
           });
         } else if (
-          !this._isReservedPasmSymbols(newName) &&
-          !this._isPasmInstruction(newName) &&
-          !this._isPasmConditional(newName) &&
-          !this._isDatStorageType(newName) &&
+          !this.parseUtils.isReservedPasmSymbols(newName) &&
+          !this.parseUtils.isPasmInstruction(newName) &&
+          !this.parseUtils.isPasmConditional(newName) &&
+          !this.parseUtils.isDatStorageType(newName) &&
           !newName.toUpperCase().startsWith("IF_")
         ) {
           this._logDAT("  --  DAT rDdl MISSING name=[" + newName + "]");
@@ -1302,14 +1501,14 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     this._logDAT("- process ValueDeclaration line(" + (lineNumber + 1) + "): line=[" + line + "]: startingOffset=(" + startingOffset + ")");
 
     // process data declaration
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const dataValueInitStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     this._logDAT("- ln(" + (lineNumber + 1) + "): dataValueInitStr=[" + dataValueInitStr + "]");
     if (dataValueInitStr.length > 1) {
       this._logDAT("  -- reportDataValueInit dataValueInitStr=[" + dataValueInitStr + "]");
 
-      let lineParts: string[] = this._getNonWhiteDataInitLineParts(dataValueInitStr);
-      const argumentStartIndex: number = this._isDatStorageType(lineParts[0]) ? 1 : 0;
+      let lineParts: string[] = this.parseUtils.getNonWhiteDataInitLineParts(dataValueInitStr);
+      const argumentStartIndex: number = this.parseUtils.isDatStorageType(lineParts[0]) ? 1 : 0;
       this._logDAT("  -- lineParts=[" + lineParts + "]");
 
       // process remainder of line
@@ -1370,13 +1569,13 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               });
             } else {
               if (
-                !this._isPasmReservedWord(namePart) &&
-                !this._isReservedPasmSymbols(namePart) &&
-                !this._isPasmInstruction(namePart) &&
-                !this._isDatNFileStorageType(namePart) &&
-                !this._isBinaryOperator(namePart) &&
-                !this._isUnaryOperator(namePart) &&
-                !this._isBuiltinReservedWord(namePart)
+                !this.parseUtils.isPasmReservedWord(namePart) &&
+                !this.parseUtils.isReservedPasmSymbols(namePart) &&
+                !this.parseUtils.isPasmInstruction(namePart) &&
+                !this.parseUtils.isDatNFileStorageType(namePart) &&
+                !this.parseUtils.isBinaryOperator(namePart) &&
+                !this.parseUtils.isUnaryOperator(namePart) &&
+                !this.parseUtils.isBuiltinReservedWord(namePart)
               ) {
                 if (showDebug) {
                   this._logMessage("  --  DAT rDvdc MISSING name=[" + namePart + "]");
@@ -1418,15 +1617,15 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportDAT_PasmCode(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const inLinePasmRHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    const lineParts: string[] = this._getNonWhitePasmLineParts(inLinePasmRHSStr);
+    const lineParts: string[] = this.parseUtils.getNonWhitePasmLineParts(inLinePasmRHSStr);
     currentOffset = line.indexOf(inLinePasmRHSStr, currentOffset);
     this._logPASM("  -- reportDATPasmDecl lineParts=[" + lineParts + "]");
     // handle name in 1 column
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     // TODO: REWRITE this to handle "non-label" line with unknown op-code!
     if (haveLabel) {
       // YES Label
@@ -1494,7 +1693,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               // skip empty operand
               continue;
             }
-            if (index == lineParts.length - 1 && this._isPasmConditional(argumentName)) {
+            if (index == lineParts.length - 1 && this.parseUtils.isPasmConditional(argumentName)) {
               // conditional flag-set spec.
               this._logPASM("  -- SKIP argumentName=[" + argumentName + "]");
               continue;
@@ -1542,11 +1741,11 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 });
               } else {
                 if (
-                  !this._isPasmReservedWord(namePart) &&
-                  !this._isPasmInstruction(namePart) &&
-                  !this._isPasmConditional(namePart) &&
-                  !this._isBinaryOperator(namePart) &&
-                  !this._isBuiltinReservedWord(namePart)
+                  !this.parseUtils.isPasmReservedWord(namePart) &&
+                  !this.parseUtils.isPasmInstruction(namePart) &&
+                  !this.parseUtils.isPasmConditional(namePart) &&
+                  !this.parseUtils.isBinaryOperator(namePart) &&
+                  !this.parseUtils.isBuiltinReservedWord(namePart)
                 ) {
                   this._logPASM("  --  DAT Pasm MISSING name=[" + namePart + "](" + (nameOffset + 1) + ")");
                   tokenSet.push({
@@ -1577,7 +1776,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             currentOffset += currArgumentLen + 1;
           }
         }
-      } else if (this._isSpin2ReservedWords(lineParts[0])) {
+      } else if (this.parseUtils.isSpin2ReservedWords(lineParts[0])) {
         const namePart: string = lineParts[argumentOffset];
         let nameOffset: number = line.indexOf(namePart, currentOffset);
         this._logPASM("  --  DAT Pasm ILLEGAL use of Pasm2 name=[" + namePart + "](" + (nameOffset + 1) + ")");
@@ -1591,7 +1790,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       }
     } else {
       // process data declaration
-      if (this._isDatStorageType(lineParts[0])) {
+      if (this.parseUtils.isDatStorageType(lineParts[0])) {
         currentOffset = line.indexOf(lineParts[0], currentOffset);
       } else {
         // skip line part 0 length when searching for [1] name
@@ -1615,7 +1814,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     this.localTokens.clear();
     const isPrivate = methodType.indexOf("PRI") != -1;
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const spineDeclarationLHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     if (spineDeclarationLHSStr) {
     } // we don't use this string, we called this to record our rhs comment!
@@ -1789,7 +1988,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     ptTokenModifiers: referenceDetails.modifiers,
                   });
                 } else {
-                  if (!this._isSpinReservedWord(namedIndexPart) && !this._isSpinBuiltinMethod(namedIndexPart) && !this._isBuiltinReservedWord(namedIndexPart)) {
+                  if (!this.parseUtils.isSpinReservedWord(namedIndexPart) && !this.parseUtils.isSpinBuiltinMethod(namedIndexPart) && !this.parseUtils.isBuiltinReservedWord(namedIndexPart)) {
                     // we don't have name registered so just mark it
                     this._logSPIN("  --  SPIN MISSING varname=[" + namedIndexPart + "](" + (nameOffset + 1) + ")");
                     tokenSet.push({
@@ -1819,7 +2018,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             this._setLocalToken(localName, new RememberedToken("variable", ["local"])); // TOKEN SET in _report()
           } else {
             // have modifier!
-            if (this._isStorageType(localName)) {
+            if (this.parseUtils.isStorageType(localName)) {
               tokenSet.push({
                 line: lineNumber,
                 startCharacter: nameOffset,
@@ -1839,7 +2038,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportSPIN_Code(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const nonCommentSpinLine = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     const remainingLength: number = nonCommentSpinLine.length;
     this._logCON("- reportSPIN nonCommentSpinLine=[" + nonCommentSpinLine + "] remainingLength=" + remainingLength);
@@ -1890,7 +2089,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               if (variableNamePart.substr(0, 1).match(/[a-zA-Z_]/)) {
                 if (variableNamePart.includes(".")) {
                   const varNameParts: string[] = variableNamePart.split(".");
-                  if (this._isDatStorageType(varNameParts[1])) {
+                  if (this.parseUtils.isDatStorageType(varNameParts[1])) {
                     variableNamePart = varNameParts[0]; // just use first part of name
                     /*
                                         // FIXME: UNDONE mark storage part correctly, yes, out-of-order
@@ -1907,7 +2106,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   }
                 }
                 this._logSPIN("  -- variableNamePart=[" + variableNamePart + "](" + (nameOffset + 1) + ")");
-                if (this._isStorageType(variableNamePart)) {
+                if (this.parseUtils.isStorageType(variableNamePart)) {
                   tokenSet.push({
                     line: lineNumber,
                     startCharacter: nameOffset,
@@ -1935,7 +2134,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                       ptTokenModifiers: modificationArray,
                     });
                   } else {
-                    if (!this._isSpinReservedWord(variableNamePart) && !this._isBuiltinReservedWord(variableNamePart) && !this._isSpinBuiltinMethod(variableNamePart)) {
+                    if (!this.parseUtils.isSpinReservedWord(variableNamePart) && !this.parseUtils.isBuiltinReservedWord(variableNamePart) && !this.parseUtils.isSpinBuiltinMethod(variableNamePart)) {
                       // we don't have name registered so just mark it
                       this._logSPIN("  --  SPIN MISSING varname=[" + variableNamePart + "](" + (nameOffset + 1) + ")");
                       tokenSet.push({
@@ -1954,11 +2153,11 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             // have simple target name, no []
             let cleanedVariableName: string = variableName.replace(/[ \t\(\)]/, "");
             const nameOffset = line.indexOf(cleanedVariableName, currentOffset);
-            if (cleanedVariableName.substr(0, 1).match(/[a-zA-Z_]/) && !this._isStorageType(cleanedVariableName)) {
+            if (cleanedVariableName.substr(0, 1).match(/[a-zA-Z_]/) && !this.parseUtils.isStorageType(cleanedVariableName)) {
               this._logSPIN("  --  SPIN cleanedVariableName=[" + cleanedVariableName + "](" + (nameOffset + 1) + ")");
               if (cleanedVariableName.includes(".")) {
                 const varNameParts: string[] = cleanedVariableName.split(".");
-                if (this._isDatStorageType(varNameParts[1])) {
+                if (this.parseUtils.isDatStorageType(varNameParts[1])) {
                   cleanedVariableName = varNameParts[0]; // just use first part of name
                   /*
                                     // FIXME: UNDONE mark storage part correctly, yes, out-of-order
@@ -2012,7 +2211,11 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 });
               } else {
                 // we don't have name registered so just mark it
-                if (!this._isSpinReservedWord(cleanedVariableName) && !this._isSpinBuiltinMethod(cleanedVariableName) && !this._isBuiltinReservedWord(cleanedVariableName)) {
+                if (
+                  !this.parseUtils.isSpinReservedWord(cleanedVariableName) &&
+                  !this.parseUtils.isSpinBuiltinMethod(cleanedVariableName) &&
+                  !this.parseUtils.isBuiltinReservedWord(cleanedVariableName)
+                ) {
                   this._logSPIN("  --  SPIN MISSING cln name=[" + cleanedVariableName + "](" + (nameOffset + 1) + ")");
                   tokenSet.push({
                     line: lineNumber,
@@ -2035,7 +2238,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       // -------------------------------------------
       const assignmentRHSStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
       currentOffset = line.indexOf(assignmentRHSStr, currentOffset);
-      const preCleanAssignmentRHSStr = this._getNonInlineCommentLine(assignmentRHSStr).replace("..", "  ");
+      const preCleanAssignmentRHSStr = this.parseUtils.getNonInlineCommentLine(assignmentRHSStr).replace("..", "  ");
       this._logSPIN("  -- SPIN assignmentRHSStr=[" + assignmentRHSStr + "]");
       const lineInfo: IFilteredStrings = this._getNonWhiteSpinLineParts(preCleanAssignmentRHSStr);
       let possNames: string[] = lineInfo.lineParts;
@@ -2098,7 +2301,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             });
           } else {
             // have unknown name!? is storage type spec?
-            if (this._isStorageType(namePart)) {
+            if (this.parseUtils.isStorageType(namePart)) {
               this._logSPIN("  --  SPIN RHS storageType=[" + namePart + "]");
               tokenSet.push({
                 line: lineNumber,
@@ -2107,7 +2310,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 ptTokenType: "storageType",
                 ptTokenModifiers: [],
               });
-            } else if (this._isSpinBuiltInConstant(namePart)) {
+            } else if (this.parseUtils.isSpinBuiltInConstant(namePart)) {
               this._logSPIN("  --  SPIN RHS builtin constant=[" + namePart + "]");
               tokenSet.push({
                 line: lineNumber,
@@ -2125,7 +2328,12 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 ptTokenType: "returnValue",
                 ptTokenModifiers: ["declaration", "local"],
               });
-            } else if (!this._isSpinReservedWord(namePart) && !this._isSpinBuiltinMethod(namePart) && !this._isSpinBuiltInVariable(namePart) && !this._isBuiltinReservedWord(namePart)) {
+            } else if (
+              !this.parseUtils.isSpinReservedWord(namePart) &&
+              !this.parseUtils.isSpinBuiltinMethod(namePart) &&
+              !this.parseUtils.isSpinBuiltInVariable(namePart) &&
+              !this.parseUtils.isBuiltinReservedWord(namePart)
+            ) {
               // NO DEBUG FOR ELSE, most of spin control elements come through here!
               //else {
               //    this._logSPIN('  -- UNKNOWN?? name=[' + namePart + '] - name-get-breakage??');
@@ -2155,7 +2363,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
               isMethod = true;
             }
             const constantPart: string = possibleNameSet[1];
-            if (this._isStorageType(constantPart)) {
+            if (this.parseUtils.isStorageType(constantPart)) {
               // FIXME: UNDONE remove when syntax see this correctly
               const nameOffset: number = line.indexOf(constantPart, currentOffset);
               this._logSPIN("  --  SPIN rhs storageType=[" + constantPart + "](" + (nameOffset + 1) + ")");
@@ -2201,18 +2409,18 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportSPIN_PasmCode(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const inLinePasmRHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    const lineParts: string[] = this._getNonWhitePasmLineParts(inLinePasmRHSStr);
+    const lineParts: string[] = this.parseUtils.getNonWhitePasmLineParts(inLinePasmRHSStr);
     this._logPASM("  -- reportInLinePasmDecl lineParts=[" + lineParts + "]");
     if (lineParts.length == 0) {
       return tokenSet;
     }
     // handle name in as first part of line...
     // (process label/variable name)
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
       this._logPASM("  -- labelName=[" + labelName + "]");
@@ -2267,7 +2475,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 // skip empty operand
                 continue;
               }
-              if (index == lineParts.length - 1 && this._isPasmConditional(argumentName)) {
+              if (index == lineParts.length - 1 && this.parseUtils.isPasmConditional(argumentName)) {
                 // conditional flag-set spec.
                 this._logPASM("  -- SKIP argumentName=[" + argumentName + "]");
                 continue;
@@ -2316,7 +2524,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   // we don't have name registered so just mark it
                   if (namePart != ".") {
                     // odd special case!
-                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart)) {
+                    if (!this.parseUtils.isSpinReservedWord(namePart) && !this.parseUtils.isBuiltinReservedWord(namePart)) {
                       this._logPASM("  --  SPIN Pasm MISSING name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -2325,7 +2533,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                         ptTokenType: "variable",
                         ptTokenModifiers: ["missingDeclaration"],
                       });
-                    } else if (this._isIllegalInlinePasmDirective(namePart)) {
+                    } else if (this.parseUtils.isIllegalInlinePasmDirective(namePart)) {
                       this._logPASM("  --  SPIN inlinePasm ERROR[CODE] illegal name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -2363,7 +2571,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
         const nameOrDirective: string = lineParts[0];
         // if this symbol is NOT a global token then it could be bad!
         if (!this._isKnownToken(nameOrDirective)) {
-          if (this._isIllegalInlinePasmDirective(nameOrDirective) || !this._isPasmInstruction(nameOrDirective)) {
+          if (this.parseUtils.isIllegalInlinePasmDirective(nameOrDirective) || !this.parseUtils.isPasmInstruction(nameOrDirective)) {
             this._logPASM("  --  SPIN inline-Pasm MISSING name=[" + nameOrDirective + "]");
             const nameOffset = line.indexOf(nameOrDirective, currentOffset);
             tokenSet.push({
@@ -2378,7 +2586,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       }
     } else {
       // process data declaration
-      if (this._isDatStorageType(lineParts[0])) {
+      if (this.parseUtils.isDatStorageType(lineParts[0])) {
         currentOffset = line.indexOf(lineParts[0], currentOffset);
       } else {
         currentOffset = line.indexOf(lineParts[1], currentOffset);
@@ -2396,7 +2604,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportOBJ_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const remainingNonCommentLineStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     //this._logOBJ('- RptObjDecl remainingNonCommentLineStr=[' + remainingNonCommentLineStr + ']');
     const remainingLength: number = remainingNonCommentLineStr.length;
@@ -2461,7 +2669,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   });
                 }
                 // we don't have name registered so just mark it
-                else if (!this._isSpinReservedWord(nameReference) && !this._isBuiltinReservedWord(nameReference)) {
+                else if (!this.parseUtils.isSpinReservedWord(nameReference) && !this.parseUtils.isBuiltinReservedWord(nameReference)) {
                   this._logOBJ("  --  OBJ MISSING name=[" + nameReference + "]");
                   tokenSet.push({
                     line: lineNumber,
@@ -2483,15 +2691,15 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportVAR_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const remainingNonCommentLineStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     if (remainingNonCommentLineStr.length > 0) {
       // get line parts - we only care about first one
-      let lineParts: string[] = this._getCommaDelimitedNonWhiteLineParts(remainingNonCommentLineStr);
+      let lineParts: string[] = this.parseUtils.getCommaDelimitedNonWhiteLineParts(remainingNonCommentLineStr);
       this._logVAR("  -- rptVarDecl lineParts=[" + lineParts + "]");
       // remember this object name so we can annotate a call to it
       const isMultiDeclaration: boolean = remainingNonCommentLineStr.includes(",");
-      const hasStorageType: boolean = this._isStorageType(lineParts[0]);
+      const hasStorageType: boolean = this.parseUtils.isStorageType(lineParts[0]);
       if (lineParts.length > 1) {
         const startIndex: number = hasStorageType ? 1 : 0;
         for (let index = startIndex; index < lineParts.length; index++) {
@@ -2551,7 +2759,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     });
                   } else {
                     // we don't have name registered so just mark it
-                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart)) {
+                    if (!this.parseUtils.isSpinReservedWord(namePart) && !this.parseUtils.isBuiltinReservedWord(namePart)) {
                       this._logVAR("  --  VAR Add MISSING name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -2632,7 +2840,7 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
             isMethod = true;
           }
           const constantPart: string = possibleNameSet[1];
-          if (this._isStorageType(constantPart)) {
+          if (this.parseUtils.isStorageType(constantPart)) {
             // FIXME: UNDONE remove when syntax see this correctly
             const nameOffset: number = line.indexOf(constantPart, referenceOffset + refPart.length + 1);
             this._logSPIN("  --  rOr rhs storageType=[" + constantPart + "](" + (nameOffset + 1) + ")");
@@ -2813,41 +3021,23 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
     };
   }
 
-  private _getNonInlineCommentLine(line: string): string {
-    // NEW remove {comment} and {{comment}} single-line elements too
-    let nonInlineCommentStr: string = line;
-    // TODO: UNDONE make this into loop to find all single line {} or {{}} comments
-    const startDoubleBraceOffset: number = nonInlineCommentStr.indexOf("{{");
-    if (startDoubleBraceOffset != -1) {
-      const endDoubleBraceOffset: number = nonInlineCommentStr.indexOf("}}", startDoubleBraceOffset + 2);
-      if (endDoubleBraceOffset != -1) {
-        // remove this comment
-        const badElement = nonInlineCommentStr.substr(startDoubleBraceOffset, endDoubleBraceOffset - startDoubleBraceOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        nonInlineCommentStr = nonInlineCommentStr.replace(badElement, " ".repeat(badElement.length));
-      }
+  private _getNonWhiteSpinLineParts(line: string): IFilteredStrings {
+    //                                     split(/[ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\#\^\/]/);
+    const nonEqualsLine: string = this.parseUtils.removeDoubleQuotedStrings(line);
+    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\^\/]+/g);
+    if (lineParts === null) {
+      lineParts = [];
     }
-    const startSingleBraceOffset: number = nonInlineCommentStr.indexOf("{");
-    if (startSingleBraceOffset != -1) {
-      const endSingleBraceOffset: number = nonInlineCommentStr.indexOf("}", startSingleBraceOffset + 1);
-      if (endSingleBraceOffset != -1) {
-        // remove this comment
-        const badElement = nonInlineCommentStr.substr(startSingleBraceOffset, endSingleBraceOffset - startSingleBraceOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        nonInlineCommentStr = nonInlineCommentStr.replace(badElement, " ".repeat(badElement.length));
-      }
-    }
-    //if (nonInlineCommentStr.length != line.length) {
-    //    this._logMessage('  -- NIC line [' + line + ']');
-    //    this._logMessage('  --          [' + nonInlineCommentStr + ']');
-    //}
-    return nonInlineCommentStr;
+    return {
+      lineNoQuotes: nonEqualsLine,
+      lineParts: lineParts,
+    };
   }
 
   private _getNonCommentLineReturnComment(lineNumber: number, startingOffset: number, line: string, tokenSet: IParsedToken[]): string {
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const nonCommentLHSStr = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const nonCommentLHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     // now record the comment if we have one
     const commentRHSStrOffset: number = currentOffset + nonCommentLHSStr.length;
     const commentOffset: number = line.indexOf("'", commentRHSStrOffset);
@@ -2863,182 +3053,6 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       tokenSet.push(newToken);
     }
     return nonCommentLHSStr;
-  }
-
-  private _getNonCommentLineRemainder(startingOffset: number, line: string): string {
-    let nonCommentRHSStr: string = line;
-    //this._logMessage('  -- gnclr ofs=' + startingOffset + '[' + line + '](' + line.length + ')');
-    // TODO: UNDONE make this into loop to find first ' not in string
-    if (line.length - startingOffset > 0) {
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], startingOffset=[' + line + ']');
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      // get line parts - we only care about first one
-      let beginCommentOffset: number = line.indexOf("'", currentOffset);
-      if (beginCommentOffset != -1) {
-        // have single quote, is it within quoted string?
-        const startDoubleQuoteOffset: number = line.indexOf('"', currentOffset);
-        if (startDoubleQuoteOffset != -1) {
-          const nonStringLine: string = this._removeDoubleQuotedStrings(line, false); // false disabled debug output
-          beginCommentOffset = nonStringLine.indexOf("'", currentOffset);
-        }
-      }
-      if (beginCommentOffset === -1) {
-        beginCommentOffset = line.indexOf("{", currentOffset);
-      }
-      const nonCommentEOL: number = beginCommentOffset != -1 ? beginCommentOffset - 1 : line.length - 1;
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], currentOffset=[' + currentOffset + ']');
-      nonCommentRHSStr = line.substr(currentOffset, nonCommentEOL - currentOffset + 1).trim();
-      //this._logMessage('- gnclr nonCommentRHSStr=[' + startingOffset + ']');
-
-      const singleLineMultiBeginOffset: number = nonCommentRHSStr.indexOf("{", currentOffset);
-      if (singleLineMultiBeginOffset != -1) {
-        const singleLineMultiEndOffset: number = nonCommentRHSStr.indexOf("}", singleLineMultiBeginOffset);
-        if (singleLineMultiEndOffset != -1) {
-          const oneLineMultiComment: string = nonCommentRHSStr.substr(singleLineMultiBeginOffset, singleLineMultiEndOffset - singleLineMultiBeginOffset + 1);
-          nonCommentRHSStr = nonCommentRHSStr.replace(oneLineMultiComment, "").trim();
-        }
-      }
-    } else if (line.length - startingOffset == 0) {
-      nonCommentRHSStr = "";
-    }
-    //if (line.substr(startingOffset).length != nonCommentRHSStr.length) {
-    //    this._logMessage('  -- NCLR line [' + line.substr(startingOffset) + ']');
-    //    this._logMessage('  --           [' + nonCommentRHSStr + ']');
-    //}
-    return nonCommentRHSStr;
-  }
-
-  private _indexOfMatchingCloseParen(line: string, openParenOffset: number): number {
-    let desiredCloseOffset: number = -1;
-    let nestingDepth: number = 1;
-    for (let offset = openParenOffset + 1; offset < line.length; offset++) {
-      if (line.substring(offset, offset + 1) == "(") {
-        nestingDepth++;
-      } else if (line.substring(offset, offset + 1) == ")") {
-        nestingDepth--;
-        if (nestingDepth == 0) {
-          // we closed the inital open
-          desiredCloseOffset = offset;
-          break; // done, get outta here
-        }
-      }
-    }
-    // this._logMessage('  -- iomcp line=[' + line + ']');
-    // this._logMessage('  --       open=(' + openParenOffset + '), close=(' + desiredCloseOffset + ')');
-    return desiredCloseOffset;
-  }
-
-  private _getNonDocCommentLineRemainder(startingOffset: number, line: string): string {
-    let nonDocCommentRHSStr: string = line;
-    //this._logMessage('  -- gnclr ofs=' + startingOffset + '[' + line + '](' + line.length + ')');
-    // TODO: UNDONE make this into loop to find first ' not in string
-    if (line.length - startingOffset > 0) {
-      const nonCommentEOL: number = line.length - 1;
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], currentOffset=[' + currentOffset + ']');
-      nonDocCommentRHSStr = line.substr(startingOffset, nonCommentEOL - startingOffset + 1).trim();
-      //this._logMessage('- gnclr nonCommentRHSStr=[' + startingOffset + ']');
-
-      const singleLineMultiBeginOffset: number = nonDocCommentRHSStr.indexOf("{", startingOffset);
-      if (singleLineMultiBeginOffset != -1) {
-        const singleLineMultiEndOffset: number = nonDocCommentRHSStr.indexOf("}", singleLineMultiBeginOffset);
-        if (singleLineMultiEndOffset != -1) {
-          const oneLineMultiComment: string = nonDocCommentRHSStr.substr(singleLineMultiBeginOffset, singleLineMultiEndOffset - singleLineMultiBeginOffset + 1);
-          nonDocCommentRHSStr = nonDocCommentRHSStr.replace(oneLineMultiComment, "").trim();
-        }
-      }
-    } else if (line.length - startingOffset == 0) {
-      nonDocCommentRHSStr = "";
-    }
-    //if (line.substr(startingOffset).length != nonCommentRHSStr.length) {
-    //    this._logMessage('  -- NCLR line [' + line.substr(startingOffset) + ']');
-    //    this._logMessage('  --           [' + nonCommentRHSStr + ']');
-    //}
-    return nonDocCommentRHSStr;
-  }
-
-  private _getNonWhiteDataInitLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\,\[\]\(\)\+\-\/\<\>\|\*\@]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getNonWhiteCONLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^  \t\(\)\*\+\-\/\>\<\=]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getNonWhitePasmLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\,\(\)\[\]\<\>\=\?\!\^\+\*\&\|\-\\\#\@\/]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getNonWhiteSpinLineParts(line: string): IFilteredStrings {
-    //                                     split(/[ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\#\^\/]/);
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\^\/]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return {
-      lineNoQuotes: nonEqualsLine,
-      lineParts: lineParts,
-    };
-  }
-
-  private _removeDoubleQuotedStrings(line: string, showDebug: boolean = true): string {
-    //this._logMessage('- RQS line [' + line + ']');
-    let trimmedLine: string = line;
-    //this._logMessage('- RQS line [' + line + ']');
-    const doubleQuote: string = '"';
-    let quoteStartOffset: number = 0; // value doesn't matter
-    let didRemove: boolean = false;
-    while ((quoteStartOffset = trimmedLine.indexOf(doubleQuote)) != -1) {
-      const quoteEndOffset: number = trimmedLine.indexOf(doubleQuote, quoteStartOffset + 1);
-      //this._logMessage('  -- quoteStartOffset=[' + quoteStartOffset + '] quoteEndOffset=[' + quoteEndOffset + ']');
-      if (quoteEndOffset != -1) {
-        const badElement = trimmedLine.substr(quoteStartOffset, quoteEndOffset - quoteStartOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        trimmedLine = trimmedLine.replace(badElement, "#".repeat(badElement.length));
-        didRemove = showDebug ? true : false;
-        //this._logMessage('-         post[' + trimmedLine + ']');
-      } else {
-        break; // we don't handle a single double-quote
-      }
-    }
-
-    //if (didRemove) {
-    //    this._logMessage('  -- RQS line [' + line + ']');
-    //    this._logMessage('  --          [' + trimmedLine + ']');
-    //}
-
-    return trimmedLine;
-  }
-
-  private _getNonWhiteLineParts(line: string): string[] {
-    let lineParts: string[] | null = line.match(/[^ \t]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getCommaDelimitedNonWhiteLineParts(line: string): string[] {
-    let lineParts: string[] | null = line.match(/[^ \t,]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
   }
 
   // -------------------------------------------------------------------------
@@ -3174,907 +3188,6 @@ export class Spin1DocumentSemanticTokensProvider implements vscode.DocumentSeman
       }
     }
     return updatedModifiers;
-  }
-
-  // -------------------------------------------------------------------------
-  // keyword checks
-  private _isBuiltinReservedWord(name: string): boolean {
-    // streamer constants, smart-pin constants
-    const builtinNamesOfNote: string[] = [
-      //
-      "dira",
-      "dirb",
-      "ina",
-      "inb",
-      "outa",
-      "outb",
-      "cnt",
-      "ctra",
-      "ctrb",
-      "frqa",
-      "frqb",
-      "phsa",
-      "phsb",
-      "vcfg",
-      "vscl",
-      "par",
-      "spr",
-    ];
-    const reservedStatus: boolean = builtinNamesOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinBuiltInConstant(name: string): boolean {
-    const spinVariablesOfNote: string[] = ["clkmode", "clkfreq", "chipver", "cogid", "cnt", "xtal1", "xtal2", "xtal3", "rcfast", "rcslow", "pll1x", "pll2x", "pll4x", "pll8x", "pll16x"];
-    let reservedStatus: boolean = spinVariablesOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinBuiltInVariable(name: string): boolean {
-    const spinVariablesOfNote: string[] = ["_clkmode", "_xinfreq", "dira", "dirb", "ina", "inb", "outa", "outb", "ctra", "ctrb", "frqa", "frqb", "phsa", "phsb", "vcfg", "vscl", "par", "spr"];
-    let reservedStatus: boolean = spinVariablesOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinReservedWord(name: string): boolean {
-    const spinInstructionsOfNote: string[] = [
-      "float",
-      "round",
-      "trunc",
-      "nan",
-      "_clkmode",
-      "_clkfreq",
-      "_free",
-      "_stack",
-      "if",
-      "ifnot",
-      "elseif",
-      "elseifnot",
-      "else",
-      "while",
-      "repeat",
-      "until",
-      "from",
-      "to",
-      "step",
-      "next",
-      "quit",
-      "case",
-      "other",
-      "abort",
-      "return",
-      "true",
-      "false",
-      "posx",
-      "negx",
-    ];
-    let reservedStatus: boolean = spinInstructionsOfNote.indexOf(name.toLowerCase()) != -1;
-    if (reservedStatus == false) {
-      reservedStatus = this._isBinaryOperator(name);
-    }
-    if (reservedStatus == false) {
-      reservedStatus = this._isUnaryOperator(name);
-    }
-    return reservedStatus;
-  }
-
-  private _isSpin2ReservedWords(name: string): boolean {
-    const spin2InstructionsOfNote: string[] = [
-      "alignl",
-      "alignw",
-      "orgf",
-      "orgh",
-      "pr0",
-      "pr1",
-      "pr2",
-      "pr3",
-      "pr4",
-      "pr5",
-      "pr6",
-      "pr7",
-      "ijmp1",
-      "ijmp2",
-      "ijmp3",
-      "iret1",
-      "iret2",
-      "iret3",
-      "pa",
-      "pb",
-      "ptra",
-      "ptrb",
-      "addct1",
-      "addct2",
-      "addct3",
-      "addpix",
-      "akpin",
-      "allowi",
-      "altb",
-      "altd",
-      "altgb",
-      "altgn",
-      "altgw",
-      "alti",
-      "altr",
-      "alts",
-      "altsb",
-      "altsn",
-      "altsw",
-      "augd",
-      "augs",
-      "bitc",
-      "bith",
-      "bitl",
-      "bitnc",
-      "bitnot",
-      "bitnz",
-      "bitrnd",
-      "bitz",
-      "blnpix",
-      "bmask",
-      "brk",
-      "calla",
-      "callb",
-      "calld",
-      "callpa",
-      "callpb",
-      "cmpm",
-      "cmpr",
-      "cogatn",
-      "cogbrk",
-      "crcbit",
-      "crcnib",
-      "decmod",
-      "decod",
-      "dirc",
-      "dirh",
-      "dirl",
-      "dirnc",
-      "dirnot",
-      "dirnz",
-      "dirrnd",
-      "dirz",
-      "djf",
-      "djnf",
-      "djz",
-      "drvc",
-      "drvh",
-      "drvl",
-      "drvnc",
-      "drvnot",
-      "drvnz",
-      "drvrnd",
-      "drvz",
-      "encod",
-      "execf",
-      "fblock",
-      "fge",
-      "fges",
-      "fle",
-      "fles",
-      "fltc",
-      "flth",
-      "fltl",
-      "fltnc",
-      "fltnot",
-      "fltnz",
-      "fltrnd",
-      "fltz",
-      "getbrk",
-      "getbyte",
-      "getbyte",
-      "getct",
-      "getnib",
-      "getptr",
-      "getqx",
-      "getqy",
-      "getrnd",
-      "getrnd",
-      "getscp",
-      "getword",
-      "getword",
-      "getxacc",
-      "hubset",
-      "ijnz",
-      "ijz",
-      "incmod",
-      "jatn",
-      "jct1",
-      "jct2",
-      "jct3",
-      "jfbw",
-      "jint",
-      "jmprel",
-      "jnatn",
-      "jnct1",
-      "jnct2",
-      "jnct3",
-      "jnfbw",
-      "jnint",
-      "jnpat",
-      "jnqmt",
-      "jnse1",
-      "jnse2",
-      "jnse3",
-      "jnse4",
-      "jnxfi",
-      "jnxmt",
-      "jnxrl",
-      "jnxro",
-      "jpat",
-      "jqmt",
-      "jse1",
-      "jse2",
-      "jse3",
-      "jse4",
-      "jxfi",
-      "jxmt",
-      "jxrl",
-      "jxro",
-      "loc",
-      "lockrel",
-      "locktry",
-      "mergeb",
-      "mergew",
-      "mixpix",
-      "modc",
-      "modcz",
-      "modz",
-      "movbyts",
-      "mulpix",
-      "muxnibs",
-      "muxnits",
-      "muxq",
-      "nixint1",
-      "nixint2",
-      "nixint3",
-      "outc",
-      "outh",
-      "outl",
-      "outnc",
-      "outnot",
-      "outnz",
-      "outrnd",
-      "outz",
-      "pollatn",
-      "pollct1",
-      "pollct2",
-      "pollct3",
-      "pollfbw",
-      "pollint",
-      "pollpat",
-      "pollqmt",
-      "pollse1",
-      "pollse2",
-      "pollse3",
-      "pollse4",
-      "pollxfi",
-      "pollxmt",
-      "pollxrl",
-      "pollxro",
-      "pop",
-      "popa",
-      "popb",
-      "push",
-      "pusha",
-      "pushb",
-      "qdiv",
-      "qexp",
-      "qfrac",
-      "qlog",
-      "qmul",
-      "qrotate",
-      "qsqrt",
-      "qvector",
-      "rczl",
-      "rczr",
-      "rdfast",
-      "rdlut",
-      "rdpin",
-      "rep",
-      "resi0",
-      "resi1",
-      "resi2",
-      "resi3",
-      "reta",
-      "retb",
-      "reti0",
-      "reti1",
-      "reti2",
-      "reti3",
-      "rfbyte",
-      "rflong",
-      "rfvar",
-      "rfvars",
-      "rfword",
-      "rgbexp",
-      "rgbsqz",
-      "rolbyte",
-      "rolbyte",
-      "rolnib",
-      "rolword",
-      "rolword",
-      "rqpin",
-      "sal",
-      "sca",
-      "scas",
-      "setbyte",
-      "setcfrq",
-      "setci",
-      "setcmod",
-      "setcq",
-      "setcy",
-      "setd",
-      "setdacs",
-      "setint1",
-      "setint2",
-      "setint3",
-      "setluts",
-      "setnib",
-      "setpat",
-      "setpiv",
-      "setpix",
-      "setq",
-      "setq2",
-      "setr",
-      "sets",
-      "setscp",
-      "setse1",
-      "setse2",
-      "setse3",
-      "setse4",
-      "setword",
-      "setxfrq",
-      "seussf",
-      "seussr",
-      "signx",
-      "skip",
-      "skipf",
-      "splitb",
-      "splitw",
-      "stalli",
-      "subr",
-      "testb",
-      "testbn",
-      "testp",
-      "testpn",
-      "tjf",
-      "tjnf",
-      "tjns",
-      "",
-      "tjs",
-      "tjv",
-      "trgint1",
-      "trgint2",
-      "trgint3",
-      "waitatn",
-      "waitct1",
-      "waitct2",
-      "waitct3",
-      "waitfbw",
-      "waitint",
-      "waitpat",
-      "waitse1",
-      "waitse2",
-      "waitse3",
-      "waitse4",
-      "waitx",
-      "waitxfi",
-      "waitxmt",
-      "waitxrl",
-      "waitxro",
-      "wfbyte",
-      "wflong",
-      "wfword",
-      "wmlong",
-      "wrc",
-      "wrfast",
-      "wrlut",
-      "wrnc",
-      "wrnz",
-      "wrpin",
-      "wrz",
-      "wxpin",
-      "wypin",
-      "xcont",
-      "xinit",
-      "xoro32",
-      "xstop",
-      "xzero",
-      "zerox",
-      "wcz",
-      "xorc",
-      "xorz",
-      "orc",
-      "orz",
-      "andc",
-      "andz",
-      "_ret_",
-      "if_00",
-      "if_01",
-      "if_0x",
-      "if_10",
-      "if_x0",
-      "if_diff",
-      "if_not_11",
-      "if_11",
-      "if_same",
-      "if_x1",
-      "if_not_10",
-      "if_1x",
-      "if_not_01",
-      "if_not_00",
-      "if_le",
-      "if_lt",
-      "if_ge",
-      "if_gt",
-      "p_adc",
-      "p_adc_100x",
-      "p_adc_10x",
-      "p_adc_1x",
-      "p_adc_30x",
-      "p_adc_3x",
-      "p_adc_ext",
-      "p_adc_float",
-      "p_adc_gio",
-      "p_adc_scope",
-      "p_adc_vio",
-      "p_async_io",
-      "p_async_rx",
-      "p_async_tx",
-      "p_bitdac",
-      "p_channel",
-      "p_compare_ab",
-      "p_compare_ab_fb",
-      "p_counter_highs",
-      "p_counter_periods",
-      "p_counter_ticks",
-      "p_count_highs",
-      "p_count_rises",
-      "p_dac_124r_3v",
-      "p_dac_600r_2v",
-      "p_dac_75r_2v",
-      "p_dac_990r_3v",
-      "p_dac_dither_pwm",
-      "p_dac_dither_rnd",
-      "p_dac_noise",
-      "p_events_ticks",
-      "p_high_100ua",
-      "p_high_10ua",
-      "p_high_150k",
-      "p_high_15k",
-      "p_high_1k5",
-      "p_high_1ma",
-      "p_high_fast",
-      "p_high_float",
-      "p_high_ticks",
-      "p_invert_a",
-      "p_invert_b",
-      "p_invert_in",
-      "p_invert_output",
-      "p_level_a",
-      "p_level_a_fbn",
-      "p_level_a_fbp",
-      "p_local_a",
-      "p_local_b",
-      "p_logic_a",
-      "p_logic_a_fb",
-      "p_logic_b_fb",
-      "p_low_100ua",
-      "p_low_10ua",
-      "p_low_150k",
-      "p_low_15k",
-      "p_low_1k5",
-      "p_low_1ma",
-      "p_low_fast",
-      "p_low_float",
-      "p_minus1_a",
-      "p_minus1_b",
-      "p_minus2_a",
-      "p_minus2_b",
-      "p_minus3_a",
-      "p_minus3_b",
-      "p_nco_duty",
-      "p_nco_freq",
-      "p_normal",
-      "p_oe",
-      "p_outbit_a",
-      "p_outbit_b",
-      "p_periods_highs",
-      "p_periods_ticks",
-      "p_plus1_a",
-      "p_plus1_b",
-      "p_plus2_a",
-      "p_plus2_b",
-      "p_plus3_a",
-      "p_plus3_b",
-      "p_pulse",
-      "p_pwm_sawtooth",
-      "p_pwm_smps",
-      "p_pwm_triangle",
-      "p_quadrature",
-      "p_reg_down",
-      "p_reg_up",
-      "p_repository",
-      "p_schmitt_a",
-      "p_schmitt_a_fb",
-      "p_schmitt_b_fb",
-      "p_state_ticks",
-      "p_sync_io",
-      "p_sync_rx",
-      "p_sync_tx",
-      "p_transition",
-      "p_true_a",
-      "p_true_b",
-      "p_true_in",
-      "p_true_output",
-      "p_tt_00",
-      "p_tt_01",
-      "p_tt_10",
-      "p_tt_11",
-      "p_usb_pair",
-      "x_16p_2dac8_wfword",
-      "x_16p_4dac4_wfword",
-      "x_1adc8_0p_1dac8_wfbyte",
-      "x_1adc8_8p_2dac8_wfword",
-      "x_1p_1dac1_wfbyte",
-      "x_2adc8_0p_2dac8_wfword",
-      "x_2adc8_16p_4dac8_wflong",
-      "x_2p_1dac2_wfbyte",
-      "x_2p_2dac1_wfbyte",
-      "x_32p_4dac8_wflong",
-      "x_4adc8_0p_4dac8_wflong",
-      "x_4p_1dac4_wfbyte",
-      "x_4p_2dac2_wfbyte",
-      "x_4p_4dac1_wfbyte",
-      "x_8p_1dac8_wfbyte",
-      "x_8p_2dac4_wfbyte",
-      "x_8p_4dac2_wfbyte",
-      "x_alt_off",
-      "x_alt_on",
-      "x_dacs_0n0_0n0",
-      "x_dacs_0n0_x_x",
-      "x_dacs_0_0_0_0",
-      "x_dacs_0_0_x_x",
-      "x_dacs_0_x_x_x",
-      "x_dacs_1n1_0n0",
-      "x_dacs_1_0_1_0",
-      "x_dacs_1_0_x_x",
-      "x_dacs_3_2_1_0",
-      "x_dacs_off",
-      "x_dacs_x_0_x_x",
-      "x_dacs_x_x_0n0",
-      "x_dacs_x_x_0_0",
-      "x_dacs_x_x_0_x",
-      "x_dacs_x_x_1_0",
-      "x_dacs_x_x_x_0",
-      "x_dds_goertzel_sinc1",
-      "x_dds_goertzel_sinc2",
-      "x_imm_16x2_1dac2",
-      "x_imm_16x2_2dac1",
-      "x_imm_16x2_lut",
-      "x_imm_1x32_4dac8",
-      "x_imm_2x16_2dac8",
-      "x_imm_2x16_4dac4",
-      "x_imm_32x1_1dac1",
-      "x_imm_32x1_lut",
-      "x_imm_4x8_1dac8",
-      "x_imm_4x8_2dac4",
-      "x_imm_4x8_4dac2",
-      "x_imm_4x8_lut",
-      "x_imm_8x4_1dac4",
-      "x_imm_8x4_2dac2",
-      "x_imm_8x4_4dac1",
-      "x_imm_8x4_lut",
-      "x_pins_off",
-      "x_pins_on",
-      "x_rfbyte_1p_1dac1",
-      "x_rfbyte_2p_1dac2",
-      "x_rfbyte_2p_2dac1",
-      "x_rfbyte_4p_1dac4",
-      "x_rfbyte_4p_2dac2",
-      "x_rfbyte_4p_4dac1",
-      "x_rfbyte_8p_1dac8",
-      "x_rfbyte_8p_2dac4",
-      "x_rfbyte_8p_4dac2",
-      "x_rfbyte_luma8",
-      "x_rfbyte_rgb8",
-      "x_rfbyte_rgbi8",
-      "x_rflong_16x2_lut",
-      "x_rflong_32p_4dac8",
-      "x_rflong_32x1_lut",
-      "x_rflong_4x8_lut",
-      "x_rflong_8x4_lut",
-      "x_rflong_rgb24",
-      "x_rfword_16p_2dac8",
-      "x_rfword_16p_4dac4",
-      "x_rfword_rgb16",
-      "x_write_off",
-      "x_write_on",
-    ];
-    const reservedStatus: boolean = spin2InstructionsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isBinaryOperator(name: string): boolean {
-    const binaryOperationsOfNote: string[] = ["sar", "ror", "rol", "rev", "zerox", "signx", "sca", "scas", "frac", "addbits", "addpins", "and", "or", "xor"];
-    const reservedStatus: boolean = binaryOperationsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isUnaryOperator(name: string): boolean {
-    const unaryOperationsOfNote: string[] = ["not", "abs", "fabs", "encod", "decod", "bmask", "ones", "sqrt", "fsqrt", "qlog", "qexp"];
-    const reservedStatus: boolean = unaryOperationsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFloatConversion(name: string): boolean {
-    const floatConversionOfNote: string[] = ["float", "round", "trunc"];
-    const reservedStatus: boolean = floatConversionOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinBuiltinMethod(name: string): boolean {
-    const spinMethodNames: string[] = [
-      "bytefill",
-      "bytemove",
-      "call",
-      "clkset",
-      "coginit",
-      "cogspin",
-      "cogstop",
-      "cognew",
-      "lockclr",
-      "locknew",
-      "lockret",
-      "lockset",
-      "longfill",
-      "longmove",
-      "lookdown",
-      "lookdownz",
-      "lookup",
-      "lookupz",
-      "reboot",
-      "strcomp",
-      "strsize",
-      "string",
-      "constant",
-      "waitcnt",
-      "waitpeq",
-      "waitpne",
-      "waitvid",
-      "wordfill",
-      "wordmove",
-    ];
-    const reservedStatus: boolean = spinMethodNames.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFlexspinPreprocessorDirective(name: string): boolean {
-    const flexspinDirectiveOfNote: string[] = ["#define", "#ifdef", "#ifndef", "#else", "#elseifdef", "#elseifndef", "#endif", "#error", "#include", "#warn", "#undef"];
-    const reservedStatus: boolean = flexspinDirectiveOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFlexspinReservedWord(name: string): boolean {
-    const flexspinReservedswordsOfNote: string[] = [
-      "__propeller__",
-      "__propeller2__",
-      "__p2__",
-      "__flexspin__",
-      "__spincvt__",
-      "__spin2pasm__",
-      "__spin2cpp__",
-      "__have_fcache__",
-      "__cplusplus__",
-      "__date__",
-      "__file__",
-      "__line__",
-      "__time__",
-      "__version__",
-      "__debug__",
-      "__output_asm__",
-      "__output_bytecode__",
-      "__output_c__",
-      "__output_cpp__",
-    ];
-    const reservedStatus: boolean = flexspinReservedswordsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isReservedPasmSymbols(name: string): boolean {
-    const reservedPasmSymbolNames: string[] = ["org", "fit", "end"];
-    const reservedStatus: boolean = reservedPasmSymbolNames.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isPasmReservedWord(name: string): boolean {
-    const pasmReservedswordsOfNote: string[] = ["cnt", "scr", "_clkfreq", "_clkmode", "_xinfreq", "_stack", "_free", "round", "float", "trunc", "true", "false", "negx", "pi", "posx"];
-    const reservedStatus: boolean = pasmReservedswordsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isPasmInstruction(name: string): boolean {
-    const pasmInstructions: string[] = [
-      "abs",
-      "absneg",
-      "add",
-      "addabs",
-      "adds",
-      "addsx",
-      "addx",
-      "and",
-      "andn",
-      "call",
-      "clkset",
-      "cmp",
-      "cmps",
-      "cmpsub",
-      "cmpsx",
-      "cmpx",
-      "cogid",
-      "coginit",
-      "cogstop",
-      "djnz",
-      "hubop",
-      "jmp",
-      "jmpret",
-      "lockclr",
-      "locknew",
-      "lockret",
-      "lockset",
-      "max",
-      "maxs",
-      "min",
-      "mins",
-      "mov",
-      "movd",
-      "movi",
-      "movs",
-      "muxc",
-      "muxnc",
-      "muxz",
-      "muxnz",
-      "neg",
-      "negc",
-      "negnc",
-      "negnz",
-      "negz",
-      "nop",
-      "or",
-      "rcl",
-      "rcr",
-      "rdbyte",
-      "rdlong",
-      "rdword",
-      "ret",
-      "rev",
-      "rol",
-      "ror",
-      "sar",
-      "shl",
-      "shr",
-      "sub",
-      "subabs",
-      "subs",
-      "subsx",
-      "subx",
-      "sumc",
-      "sumnc",
-      "sumnz",
-      "sumz",
-      "test",
-      "testn",
-      "tjnz",
-      "tjz",
-      "waitcnt",
-      "waitpeq",
-      "waitpne",
-      "waitvid",
-      "wrbyte",
-      "wrlong",
-      "wrword",
-      "xor",
-    ];
-    const instructionStatus: boolean = pasmInstructions.indexOf(name.toLowerCase()) != -1;
-    return instructionStatus;
-  }
-
-  private _isPasmNonArgumentInstruction(name: string): boolean {
-    const pasmNonArgumentInstructions: string[] = ["nop", "ret"];
-    const instructionStatus: boolean = pasmNonArgumentInstructions.indexOf(name.toLowerCase()) != -1;
-    return instructionStatus;
-  }
-
-  private _isIllegalInlinePasmDirective(name: string): boolean {
-    const illegalInlinePasmDirective: string[] = ["file"];
-    const illegalStatus: boolean = illegalInlinePasmDirective.indexOf(name.toLowerCase()) != -1;
-    return illegalStatus;
-  }
-
-  private _isPasmConditional(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length >= 2) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "WC" || checkType == "WZ" || checkType == "NR" || checkType == "WR") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isDatOrPasmLabel(name: string): boolean {
-    let haveLabelStatus: boolean = false;
-    if (name.length > 0) {
-      haveLabelStatus = name.substr(0, 1).match(/[a-zA-Z_\.\:]/) ? true : false;
-      if (haveLabelStatus) {
-        if (this._isDatNFileStorageType(name)) {
-          haveLabelStatus = false;
-        } else if (name.toLowerCase() == "dat") {
-          haveLabelStatus = false;
-        } else if (this._isReservedPasmSymbols(name)) {
-          haveLabelStatus = false;
-        } else if (name.toUpperCase().startsWith("IF_")) {
-          haveLabelStatus = false;
-        } else if (this._isPasmConditional(name)) {
-          haveLabelStatus = false;
-        } else if (this._isIllegalInlinePasmDirective(name)) {
-          haveLabelStatus = false;
-        } else if (this._isPasmInstruction(name)) {
-          haveLabelStatus = false;
-        } else if (this._isPasmNonArgumentInstruction(name)) {
-          haveLabelStatus = false;
-        }
-      }
-    }
-    return haveLabelStatus;
-  }
-
-  private _isDatNFileStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 2) {
-      const checkType: string = name.toUpperCase();
-      // yeah, FILE too!  (oddly enough)
-      if (checkType == "BYTE" || checkType == "WORD" || checkType == "LONG" || checkType == "RES" || checkType == "FILE") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isDatStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 2) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "BYTE" || checkType == "WORD" || checkType == "LONG" || checkType == "RES") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 3) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "BYTE" || checkType == "WORD" || checkType == "LONG") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  //
-  // ----------------------------------------------------
-
-  private _skipWhite(line: string, currentOffset: number): number {
-    let firstNonWhiteIndex: number = currentOffset;
-    for (let index = currentOffset; index < line.length; index++) {
-      if (line.substr(index, 1) != " " && line.substr(index, 1) != "\t") {
-        firstNonWhiteIndex = index;
-        break;
-      }
-    }
-    return firstNonWhiteIndex;
   }
 
   private _tokenString(aToken: IParsedToken, line: string): string {
